+++ laravel/app/Http/Controllers/backend/NewsController.php
@@ -119,13 +119,10 @@ class NewsController extends Controller
 
     private function RemoveFolderImage($rawfile)
     {
-        if($rawfile != "")
-        {
-          $rawfileArr = explode("/", $rawfile);
-          $indexFile = count($rawfileArr) - 1;
-          $indexFolder = count($rawfileArr) - 2;
-          File::delete($rawfile);
-          File::deleteDirectory(config('app.upload_news').$rawfileArr[$indexFolder]);
+        if ($rawfile != "") {
+            if (File::exists($rawfile)) {
+                File::delete($rawfile);
+            }
         }
     }
